requirements.txt.txt
streamlit
pandas
pytesseract
opencv-python
numpy
Pillow
deskew